谷粒商城基础篇完结

+++ gulimall-product/src/main/java/com/yang/gulimall/product/vo/SpuSavaVo/BaseAttrs.java
@@ -14,7 +14,7 @@ import lombok.Data;
 @Data
 public class BaseAttrs {
 
-    private int attrId;
+    private Long attrId;
     private String attrValues;
     private int showDesc;
